backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
motor>=3.3.2
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
email-validator==2.1.0
